fix(protocol): canonicalize generated declarations
